+++ react-components/src/components/Table/components/RowTable/index.tsx
@@ -2,25 +2,47 @@ import React, { memo } from 'react';
 import { flexRender, Row } from '@tanstack/react-table';
 import classNames from 'classnames';
 
+import { TableProps } from '../../model';
+import { ICellInstance } from 'tetacom/react-components';
+
 import s from '../../style.module.scss';
 
 export interface ITableRow<T> {
   row: Row<T>;
+  columns: TableProps<T>['columns'];
   isSelectedRow?: boolean;
+  onClick?: (cell: ICellInstance<T>) => void;
 }
 
-function TableRow<T>({ row, isSelectedRow = false }: ITableRow<T>) {
+function TableRow<T>({
+  row,
+  columns,
+  isSelectedRow = false,
+  onClick,
+}: ITableRow<T>) {
   const { toggleSelected, getVisibleCells } = row;
 
+  const handleClick = (columnId: string, row: T) => {
+    toggleSelected();
+
+    const column = columns.find((item) => item.name === columnId);
+    if (onClick && column) {
+      onClick({
+        row,
+        column,
+      });
+    }
+  };
+
   return (
-    <tr
-      className={classNames(isSelectedRow && s.active)}
-      onClick={() => {
-        toggleSelected();
-      }}
-    >
+    <tr className={classNames(isSelectedRow && s.active)}>
       {getVisibleCells().map((cell) => (
-        <td key={cell.id}>
+        <td
+          key={cell.id}
+          onClick={() => {
+            handleClick(cell.column.id, cell.row.original);
+          }}
+        >
           {flexRender(cell.column.columnDef.cell, cell.getContext())}
         </td>
       ))}