+++ react-components/src/components/Table/model/cell-params.ts
@@ -1,0 +1,4 @@
+export type CellParamsType = {
+  maxWidth: number | string;
+  verticalClamp: number;
+};